+++ src/main/kotlin/com/example/SistemaDeVendasDeAutomoveis/Response/ClientResponse.kt
@@ -2,7 +2,7 @@ package com.example.SistemaDeVendasDeAutomoveis.Response
 
 data class ClientResponse(
 
-    val idCliente: Int,
+    val idCliente: Long,
     val nome: String,
     val email: String,
     val telefone: String
